+++ SimCoinBack/src/main/java/com/revature/controller/ToplistController.java
@@ -7,13 +7,10 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-
 import com.revature.model.User;
-import com.revature.services.UserService;
-
+import com.revature.services.TopListService;
 import io.swagger.v3.oas.annotations.parameters.RequestBody;
 
 @RestController
@@ -21,17 +18,17 @@ import io.swagger.v3.oas.annotations.parameters.RequestBody;
 @CrossOrigin("*")
 public class ToplistController {
 	
-	private UserService us;
-
+	private TopListService tls;
+	
 	@Autowired
-	public ToplistController(UserService us) {
-		this.us = us;
+	ToplistController(TopListService tls){
+		this.tls = tls;
 	}
 	
 	@GetMapping
 	public ResponseEntity<ArrayList<User>> getTop() {
 		
-		ArrayList<User> topList = us.getTop();
+		ArrayList<User> topList = tls.getTop();
 		
 		return new ResponseEntity<ArrayList<User>>(topList, HttpStatus.CREATED);
 	}
